mrc-weight-update-export(weight-updater-methods-move,mechanical_provable): Move init/destroy weights_update_group onto WeightUpdater (cut+paste)
